requirements.txt
Django==5.0.7
django-appconf==1.0.6
django-ckeditor==6.7.0
django-compressor==4.4
django-cors-headers==3.13.0
django-dbbackup==4.0.2
django-debug-toolbar==4.2.0
django-filter==22.1
django-htmlmin==0.11.0
django-import-export==3.3.1
django-jalali==6.0.0
django-jalali-date==1.0.2
django-js-asset==2.1.0
et-xmlfile==1.1.0
exceptiongroup==1.0.0
filelock==3.8.0
fontawesomefree==6.4.0
html5lib==1.1
idna==3.4
isodate==0.6.1
itypes==1.2.0
jdatetime==4.1.0
Jinja2==3.1.2
jsmin==3.0.1
lxml==4.9.3
Markdown==3.4.1
MarkupPy==1.14
MarkupSafe==2.1.1
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.1.2
outcome==1.2.0
packaging==21.3
Pillow==10.0.1
platformdirs==2.5.2
pycparser==2.21
pyOpenSSL==22.1.0
pyparsing==3.0.9
pypng==0.20220715.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.6
rcssmin==1.1.1
redis==4.3.4
requests==2.31.0
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.12
webencodings==0.5.1
wheel==0.43.0
whitenoise==6.6.0
wrapt==1.14.1
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
zeep==4.2.0

